requirements.txt
pyOpenSSL==16.2.0